flang: libflangrti.so should depend on libompstub.so instead of libomp.so
This commit is a clang part of a change that drops unfortunate
dependency of libflangrti.so on libomp.so.

For non-OpenMP programs (built without -fopenmp flag) frontend
driver instructs the linker to link against libompstub.so library,
which exports stubs for all OpenMP runtime library symbols. This is
because Flang runtime library contains calls to those symbols
(for OpenMP programs, final binary is linked against libomp.so
instead).

Apart form above, all of the Fortran programs are linked against
libflang.so and libflangrti.so. Unfortunately, before this change
libflangrti.so was depending on full blown OpenMP runtime library
(libomp.so).

This change creates two sets of Flang runtime shared objects with
following chains of dependencies:

1. libflang.so -> libflangrti.so -> libompstub.so
2. libflang-omp.so -> libflangrti-omp.so -> libomp.so

Note that before this commit is applied, relevant commit must be
applied on flang repository first. Otherwise all OpenMP Fortran programs
will fail to build at link time.

Signed-off-by: Paul Osmialowski <pawel.osmialowski@arm.com>